% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_function.R
\name{fit_SGNB}
\alias{fit_SGNB}
\title{Fit SGNB model}
\usage{
fit_SGNB(read_summarized_df, gene_size_ls = NULL, tol = 0.001,
  times = 200)
}
\arguments{
\item{read_summarized_df}{The returned data.frame from function \code{\link{summarize_read_single_end}} or
\code{\link{summarize_read_paired_end}}.}

\item{gene_size_ls}{The gene size file. If set it to NULL, the function will use the gene_size file saved in the
package data folder, which is for human. Default is NULL.}

\item{tol}{A numerical value to control estimation accuracy. Default is 0.001.}

\item{times}{A numerical value to control estimation loop times. Default is 200.}
}
\value{
A data.fram saving gene_id, p-value, likelihood test statistics, degree of freedom and gene expression
        for condition 0 and 1.
}
\description{
This function estiamte the mle and returns p-value and estimated gene expression
level in a data.frame.
}
\examples{
fit_SGNB(read_summarized_df)
fit_SGNB(read_summarized_df, gene_size, tol = 0.01, times = 500)

}
