% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_function.R
\name{fit_SGNB_exact}
\alias{fit_SGNB_exact}
\title{Fit exact SGNB model}
\usage{
fit_SGNB_exact(read_summarized_df, side = "double", min_reduce = 0,
  tol = 0.001, times = 200)
}
\arguments{
\item{read_summarized_df}{The returned data.frame from function \code{\link{summarize_read_single_end}} or
\code{\link{summarize_read_paired_end}}.}

\item{side}{test type, could be "left", "right", or "double", the default is "double"}

\item{min_reduce}{A number between 0 and 1 indicating the minimum percentage of the number of parameters
needed to be reduced. If the parameter reduction procedure fail to reduce this amount,
then group all read types into one read type. Default is 0.}

\item{tol}{A numerical value to control estimation accuracy. Default is 0.001.}

\item{times}{A numerical value to control estimation loop times. Default is 100.}
}
\value{
A data.frame having gene_id, p-value and relative gene expression for condition 0 and 1.
}
\description{
Fit exact SGNB model
}
\examples{
fit_SGNB_exact(read_summarized_df)

}
