% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_function.R
\name{modify_ann}
\alias{modify_ann}
\title{Modify gene annotation file}
\usage{
modify_ann(gene_ann_path, line_skip = 5, sep = "\\t", gene_id = "gene_id",
  block_ann_path, gene_range_path, gene_size_path)
}
\arguments{
\item{gene_ann_path}{A physical path to gene annotation file in gtf format.}

\item{line_skip}{The number of lines that need to be skipped when reading the gtf file. Default is 5.}

\item{sep}{The separation used in the gene annotation file. Default is tab.}

\item{gene_id}{A fixed start string of gene id in the gene annotation file in gtf format downloaded from Ensembl.
Default is 'gene_id'.}

\item{block_ann_path}{A physical path to save block annotation in .RData format.}

\item{gene_range_path}{A physical path to save gene range lookup table in .RData format.}

\item{gene_size_path}{A physical path to save gene size lookup table in .RData format.}
}
\value{
none.
}
\description{
This function creates 3 files in .RData format, which are block annotation,
gene range lookup table, and gene size lookup table, from the gene annotation
file in gtf format downloaded form Ensembl for later analysis.
}
\details{
The block annotation file has three levels. First is a list of chromosomes. Second is a list of
genes. Third is a data frame of block annotations.

The gene range file has two levels. First is a list of chromosomes. Second is a data frame of gene
range information.

The gene size file has one level, which is a list of gene size information.
}
\examples{
modify_ann("./Homo.gtf", 5, "./block_ann.RData", "./gene_range.RData", "./gene_size.RData")

}
