% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_function.R
\name{summarize_read_paired_end}
\alias{summarize_read_paired_end}
\title{Summarize RNA-seq paired end read}
\usage{
summarize_read_paired_end(input_sam_folder_path_0, input_sam_folder_path_1,
  block_ann_ls = NULL, gene_range_ls = NULL, run.parallel = TRUE,
  core.num = parallel::detectCores())
}
\arguments{
\item{input_sam_folder_path_0}{A physical path to a folder of sam files under condition 0.}

\item{input_sam_folder_path_1}{A physical path to a folder of sam files under condition 1.}

\item{block_ann_ls}{The block annotation file created by function \code{\link{modify_ann}}. If set to NULL,
the function will use the block_ann file saved in the package data folder, which is for
human. Default is NULL.}

\item{gene_range_ls}{The gene range file created by function \code{\link{modify_ann}}. If set to NULL,
the function will use the gene_range file saved in the package data folder, which is for
human. Default is NULL.}

\item{run.parallel}{A logical variable. If TRUE, run in parallel mode. Default is TRUE.}

\item{core.num}{CPU core numbers will be used if run.parallel is TRUE. Default is the number of all the cores.}
}
\value{
A data.frame saving the summarized read.
}
\description{
This function summarize the RNA-seq paired end read into read type.
}
\examples{
summarize_read_paired_end("./group0", "./group1")
summarize_read_paired_end("./group0", "./group1", block_ann, gene_range)

}
